chore: move file moving around
