+++ app/graphql/generated.ts
@@ -1531,86 +1531,6 @@ export type MeFragment = {
   }
 }
 
-export type DeviceNotificationTokenCreateMutationVariables = Exact<{
-  input: DeviceNotificationTokenCreateInput
-}>
-
-export type DeviceNotificationTokenCreateMutation = {
-  readonly __typename: "Mutation"
-  readonly deviceNotificationTokenCreate: {
-    readonly __typename: "SuccessPayload"
-    readonly success?: boolean | null
-    readonly errors: ReadonlyArray<{
-      readonly __typename: "GraphQLApplicationError"
-      readonly message: string
-    }>
-  }
-}
-
-export type LnNoAmountInvoiceFeeProbeMutationVariables = Exact<{
-  input: LnNoAmountInvoiceFeeProbeInput
-}>
-
-export type LnNoAmountInvoiceFeeProbeMutation = {
-  readonly __typename: "Mutation"
-  readonly lnNoAmountInvoiceFeeProbe: {
-    readonly __typename: "SatAmountPayload"
-    readonly amount?: number | null
-    readonly errors: ReadonlyArray<{
-      readonly __typename: "GraphQLApplicationError"
-      readonly message: string
-    }>
-  }
-}
-
-export type LnInvoiceFeeProbeMutationVariables = Exact<{
-  input: LnInvoiceFeeProbeInput
-}>
-
-export type LnInvoiceFeeProbeMutation = {
-  readonly __typename: "Mutation"
-  readonly lnInvoiceFeeProbe: {
-    readonly __typename: "SatAmountPayload"
-    readonly amount?: number | null
-    readonly errors: ReadonlyArray<{
-      readonly __typename: "GraphQLApplicationError"
-      readonly message: string
-    }>
-  }
-}
-
-export type LnUsdInvoiceFeeProbeMutationVariables = Exact<{
-  input: LnUsdInvoiceFeeProbeInput
-}>
-
-export type LnUsdInvoiceFeeProbeMutation = {
-  readonly __typename: "Mutation"
-  readonly lnUsdInvoiceFeeProbe: {
-    readonly __typename: "SatAmountPayload"
-    readonly amount?: number | null
-    readonly errors: ReadonlyArray<{
-      readonly __typename: "GraphQLApplicationError"
-      readonly message: string
-    }>
-  }
-}
-
-export type LnNoAmountUsdInvoiceFeeProbeMutationVariables = Exact<{
-  input: LnNoAmountUsdInvoiceFeeProbeInput
-}>
-
-export type LnNoAmountUsdInvoiceFeeProbeMutation = {
-  readonly __typename: "Mutation"
-  readonly lnNoAmountUsdInvoiceFeeProbe: {
-    readonly __typename: "CentAmountPayload"
-    readonly amount?: number | null
-    readonly errors: ReadonlyArray<{
-      readonly __typename: "GraphQLApplicationError"
-      readonly message: string
-    }>
-  }
-}
-
 export type TransactionListForContactQueryVariables = Exact<{
   username: Scalars["Username"]
   first?: InputMaybe<Scalars["Int"]>
@@ -1771,39 +1691,6 @@ export type OnChainTxFeeQuery = {
     readonly amount: number
     readonly targetConfirmations: number
   }
-}
-
-export type AccountLimitsQueryVariables = Exact<{ [key: string]: never }>
-
-export type AccountLimitsQuery = {
-  readonly __typename: "Query"
-  readonly me?: {
-    readonly __typename: "User"
-    readonly defaultAccount: {
-      readonly __typename: "ConsumerAccount"
-      readonly limits: {
-        readonly __typename: "AccountLimits"
-        readonly withdrawal: ReadonlyArray<{
-          readonly __typename: "OneDayAccountLimit"
-          readonly totalLimit: number
-          readonly remainingLimit?: number | null
-          readonly interval?: number | null
-        }>
-        readonly internalSend: ReadonlyArray<{
-          readonly __typename: "OneDayAccountLimit"
-          readonly totalLimit: number
-          readonly remainingLimit?: number | null
-          readonly interval?: number | null
-        }>
-        readonly convert: ReadonlyArray<{
-          readonly __typename: "OneDayAccountLimit"
-          readonly totalLimit: number
-          readonly remainingLimit?: number | null
-          readonly interval?: number | null
-        }>
-      }
-    }
-  } | null
 }
 
 export type WalletsQueryVariables = Exact<{ [key: string]: never }>
@@ -2628,6 +2515,70 @@ export type WalletMetaFragment =
   | WalletMeta_BtcWallet_Fragment
   | WalletMeta_UsdWallet_Fragment
 
+export type LnNoAmountInvoiceFeeProbeMutationVariables = Exact<{
+  input: LnNoAmountInvoiceFeeProbeInput
+}>
+
+export type LnNoAmountInvoiceFeeProbeMutation = {
+  readonly __typename: "Mutation"
+  readonly lnNoAmountInvoiceFeeProbe: {
+    readonly __typename: "SatAmountPayload"
+    readonly amount?: number | null
+    readonly errors: ReadonlyArray<{
+      readonly __typename: "GraphQLApplicationError"
+      readonly message: string
+    }>
+  }
+}
+
+export type LnInvoiceFeeProbeMutationVariables = Exact<{
+  input: LnInvoiceFeeProbeInput
+}>
+
+export type LnInvoiceFeeProbeMutation = {
+  readonly __typename: "Mutation"
+  readonly lnInvoiceFeeProbe: {
+    readonly __typename: "SatAmountPayload"
+    readonly amount?: number | null
+    readonly errors: ReadonlyArray<{
+      readonly __typename: "GraphQLApplicationError"
+      readonly message: string
+    }>
+  }
+}
+
+export type LnUsdInvoiceFeeProbeMutationVariables = Exact<{
+  input: LnUsdInvoiceFeeProbeInput
+}>
+
+export type LnUsdInvoiceFeeProbeMutation = {
+  readonly __typename: "Mutation"
+  readonly lnUsdInvoiceFeeProbe: {
+    readonly __typename: "SatAmountPayload"
+    readonly amount?: number | null
+    readonly errors: ReadonlyArray<{
+      readonly __typename: "GraphQLApplicationError"
+      readonly message: string
+    }>
+  }
+}
+
+export type LnNoAmountUsdInvoiceFeeProbeMutationVariables = Exact<{
+  input: LnNoAmountUsdInvoiceFeeProbeInput
+}>
+
+export type LnNoAmountUsdInvoiceFeeProbeMutation = {
+  readonly __typename: "Mutation"
+  readonly lnNoAmountUsdInvoiceFeeProbe: {
+    readonly __typename: "CentAmountPayload"
+    readonly amount?: number | null
+    readonly errors: ReadonlyArray<{
+      readonly __typename: "GraphQLApplicationError"
+      readonly message: string
+    }>
+  }
+}
+
 export type AccountScreenQueryVariables = Exact<{ [key: string]: never }>
 
 export type AccountScreenQuery = {
@@ -2702,6 +2653,39 @@ export type SettingsScreenQuery = {
   } | null
 }
 
+export type AccountLimitsQueryVariables = Exact<{ [key: string]: never }>
+
+export type AccountLimitsQuery = {
+  readonly __typename: "Query"
+  readonly me?: {
+    readonly __typename: "User"
+    readonly defaultAccount: {
+      readonly __typename: "ConsumerAccount"
+      readonly limits: {
+        readonly __typename: "AccountLimits"
+        readonly withdrawal: ReadonlyArray<{
+          readonly __typename: "OneDayAccountLimit"
+          readonly totalLimit: number
+          readonly remainingLimit?: number | null
+          readonly interval?: number | null
+        }>
+        readonly internalSend: ReadonlyArray<{
+          readonly __typename: "OneDayAccountLimit"
+          readonly totalLimit: number
+          readonly remainingLimit?: number | null
+          readonly interval?: number | null
+        }>
+        readonly convert: ReadonlyArray<{
+          readonly __typename: "OneDayAccountLimit"
+          readonly totalLimit: number
+          readonly remainingLimit?: number | null
+          readonly interval?: number | null
+        }>
+      }
+    }
+  } | null
+}
+
 export type LocalizationContextProviderQueryVariables = Exact<{ [key: string]: never }>
 
 export type LocalizationContextProviderQuery = {
@@ -2724,6 +2708,22 @@ export type PriceSubscription = {
       readonly currencyUnit: ExchangeCurrencyUnit
       readonly formattedAmount: string
     } | null
+    readonly errors: ReadonlyArray<{
+      readonly __typename: "GraphQLApplicationError"
+      readonly message: string
+    }>
+  }
+}
+
+export type DeviceNotificationTokenCreateMutationVariables = Exact<{
+  input: DeviceNotificationTokenCreateInput
+}>
+
+export type DeviceNotificationTokenCreateMutation = {
+  readonly __typename: "Mutation"
+  readonly deviceNotificationTokenCreate: {
+    readonly __typename: "SuccessPayload"
+    readonly success?: boolean | null
     readonly errors: ReadonlyArray<{
       readonly __typename: "GraphQLApplicationError"
       readonly message: string
@@ -3037,271 +3037,6 @@ export type HiddenBalanceToolTipQueryResult = Apollo.QueryResult<
   HiddenBalanceToolTipQuery,
   HiddenBalanceToolTipQueryVariables
 >
-export const DeviceNotificationTokenCreateDocument = gql`
-  mutation deviceNotificationTokenCreate($input: DeviceNotificationTokenCreateInput!) {
-    deviceNotificationTokenCreate(input: $input) {
-      errors {
-        message
-      }
-      success
-    }
-  }
-`
-export type DeviceNotificationTokenCreateMutationFn = Apollo.MutationFunction<
-  DeviceNotificationTokenCreateMutation,
-  DeviceNotificationTokenCreateMutationVariables
->
-
-/**
- * __useDeviceNotificationTokenCreateMutation__
- *
- * To run a mutation, you first call `useDeviceNotificationTokenCreateMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useDeviceNotificationTokenCreateMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [deviceNotificationTokenCreateMutation, { data, loading, error }] = useDeviceNotificationTokenCreateMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useDeviceNotificationTokenCreateMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    DeviceNotificationTokenCreateMutation,
-    DeviceNotificationTokenCreateMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<
-    DeviceNotificationTokenCreateMutation,
-    DeviceNotificationTokenCreateMutationVariables
-  >(DeviceNotificationTokenCreateDocument, options)
-}
-export type DeviceNotificationTokenCreateMutationHookResult = ReturnType<
-  typeof useDeviceNotificationTokenCreateMutation
->
-export type DeviceNotificationTokenCreateMutationResult =
-  Apollo.MutationResult<DeviceNotificationTokenCreateMutation>
-export type DeviceNotificationTokenCreateMutationOptions = Apollo.BaseMutationOptions<
-  DeviceNotificationTokenCreateMutation,
-  DeviceNotificationTokenCreateMutationVariables
->
-export const LnNoAmountInvoiceFeeProbeDocument = gql`
-  mutation lnNoAmountInvoiceFeeProbe($input: LnNoAmountInvoiceFeeProbeInput!) {
-    lnNoAmountInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-`
-export type LnNoAmountInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
-  LnNoAmountInvoiceFeeProbeMutation,
-  LnNoAmountInvoiceFeeProbeMutationVariables
->
-
-/**
- * __useLnNoAmountInvoiceFeeProbeMutation__
- *
- * To run a mutation, you first call `useLnNoAmountInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLnNoAmountInvoiceFeeProbeMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [lnNoAmountInvoiceFeeProbeMutation, { data, loading, error }] = useLnNoAmountInvoiceFeeProbeMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useLnNoAmountInvoiceFeeProbeMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    LnNoAmountInvoiceFeeProbeMutation,
-    LnNoAmountInvoiceFeeProbeMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<
-    LnNoAmountInvoiceFeeProbeMutation,
-    LnNoAmountInvoiceFeeProbeMutationVariables
-  >(LnNoAmountInvoiceFeeProbeDocument, options)
-}
-export type LnNoAmountInvoiceFeeProbeMutationHookResult = ReturnType<
-  typeof useLnNoAmountInvoiceFeeProbeMutation
->
-export type LnNoAmountInvoiceFeeProbeMutationResult =
-  Apollo.MutationResult<LnNoAmountInvoiceFeeProbeMutation>
-export type LnNoAmountInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
-  LnNoAmountInvoiceFeeProbeMutation,
-  LnNoAmountInvoiceFeeProbeMutationVariables
->
-export const LnInvoiceFeeProbeDocument = gql`
-  mutation lnInvoiceFeeProbe($input: LnInvoiceFeeProbeInput!) {
-    lnInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-`
-export type LnInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
-  LnInvoiceFeeProbeMutation,
-  LnInvoiceFeeProbeMutationVariables
->
-
-/**
- * __useLnInvoiceFeeProbeMutation__
- *
- * To run a mutation, you first call `useLnInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLnInvoiceFeeProbeMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [lnInvoiceFeeProbeMutation, { data, loading, error }] = useLnInvoiceFeeProbeMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useLnInvoiceFeeProbeMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    LnInvoiceFeeProbeMutation,
-    LnInvoiceFeeProbeMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<
-    LnInvoiceFeeProbeMutation,
-    LnInvoiceFeeProbeMutationVariables
-  >(LnInvoiceFeeProbeDocument, options)
-}
-export type LnInvoiceFeeProbeMutationHookResult = ReturnType<
-  typeof useLnInvoiceFeeProbeMutation
->
-export type LnInvoiceFeeProbeMutationResult =
-  Apollo.MutationResult<LnInvoiceFeeProbeMutation>
-export type LnInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
-  LnInvoiceFeeProbeMutation,
-  LnInvoiceFeeProbeMutationVariables
->
-export const LnUsdInvoiceFeeProbeDocument = gql`
-  mutation lnUsdInvoiceFeeProbe($input: LnUsdInvoiceFeeProbeInput!) {
-    lnUsdInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-`
-export type LnUsdInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
-  LnUsdInvoiceFeeProbeMutation,
-  LnUsdInvoiceFeeProbeMutationVariables
->
-
-/**
- * __useLnUsdInvoiceFeeProbeMutation__
- *
- * To run a mutation, you first call `useLnUsdInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLnUsdInvoiceFeeProbeMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [lnUsdInvoiceFeeProbeMutation, { data, loading, error }] = useLnUsdInvoiceFeeProbeMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useLnUsdInvoiceFeeProbeMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    LnUsdInvoiceFeeProbeMutation,
-    LnUsdInvoiceFeeProbeMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<
-    LnUsdInvoiceFeeProbeMutation,
-    LnUsdInvoiceFeeProbeMutationVariables
-  >(LnUsdInvoiceFeeProbeDocument, options)
-}
-export type LnUsdInvoiceFeeProbeMutationHookResult = ReturnType<
-  typeof useLnUsdInvoiceFeeProbeMutation
->
-export type LnUsdInvoiceFeeProbeMutationResult =
-  Apollo.MutationResult<LnUsdInvoiceFeeProbeMutation>
-export type LnUsdInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
-  LnUsdInvoiceFeeProbeMutation,
-  LnUsdInvoiceFeeProbeMutationVariables
->
-export const LnNoAmountUsdInvoiceFeeProbeDocument = gql`
-  mutation lnNoAmountUsdInvoiceFeeProbe($input: LnNoAmountUsdInvoiceFeeProbeInput!) {
-    lnNoAmountUsdInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-`
-export type LnNoAmountUsdInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
-  LnNoAmountUsdInvoiceFeeProbeMutation,
-  LnNoAmountUsdInvoiceFeeProbeMutationVariables
->
-
-/**
- * __useLnNoAmountUsdInvoiceFeeProbeMutation__
- *
- * To run a mutation, you first call `useLnNoAmountUsdInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLnNoAmountUsdInvoiceFeeProbeMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [lnNoAmountUsdInvoiceFeeProbeMutation, { data, loading, error }] = useLnNoAmountUsdInvoiceFeeProbeMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useLnNoAmountUsdInvoiceFeeProbeMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    LnNoAmountUsdInvoiceFeeProbeMutation,
-    LnNoAmountUsdInvoiceFeeProbeMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<
-    LnNoAmountUsdInvoiceFeeProbeMutation,
-    LnNoAmountUsdInvoiceFeeProbeMutationVariables
-  >(LnNoAmountUsdInvoiceFeeProbeDocument, options)
-}
-export type LnNoAmountUsdInvoiceFeeProbeMutationHookResult = ReturnType<
-  typeof useLnNoAmountUsdInvoiceFeeProbeMutation
->
-export type LnNoAmountUsdInvoiceFeeProbeMutationResult =
-  Apollo.MutationResult<LnNoAmountUsdInvoiceFeeProbeMutation>
-export type LnNoAmountUsdInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
-  LnNoAmountUsdInvoiceFeeProbeMutation,
-  LnNoAmountUsdInvoiceFeeProbeMutationVariables
->
 export const TransactionListForContactDocument = gql`
   query transactionListForContact(
     $username: Username!
@@ -3513,76 +3248,6 @@ export type OnChainTxFeeLazyQueryHookResult = ReturnType<typeof useOnChainTxFeeL
 export type OnChainTxFeeQueryResult = Apollo.QueryResult<
   OnChainTxFeeQuery,
   OnChainTxFeeQueryVariables
->
-export const AccountLimitsDocument = gql`
-  query accountLimits {
-    me {
-      defaultAccount {
-        limits {
-          withdrawal {
-            totalLimit
-            remainingLimit
-            interval
-          }
-          internalSend {
-            totalLimit
-            remainingLimit
-            interval
-          }
-          convert {
-            totalLimit
-            remainingLimit
-            interval
-          }
-        }
-      }
-    }
-  }
-`
-
-/**
- * __useAccountLimitsQuery__
- *
- * To run a query within a React component, call `useAccountLimitsQuery` and pass it any options that fit your needs.
- * When your component renders, `useAccountLimitsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useAccountLimitsQuery({
- *   variables: {
- *   },
- * });
- */
-export function useAccountLimitsQuery(
-  baseOptions?: Apollo.QueryHookOptions<AccountLimitsQuery, AccountLimitsQueryVariables>,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useQuery<AccountLimitsQuery, AccountLimitsQueryVariables>(
-    AccountLimitsDocument,
-    options,
-  )
-}
-export function useAccountLimitsLazyQuery(
-  baseOptions?: Apollo.LazyQueryHookOptions<
-    AccountLimitsQuery,
-    AccountLimitsQueryVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useLazyQuery<AccountLimitsQuery, AccountLimitsQueryVariables>(
-    AccountLimitsDocument,
-    options,
-  )
-}
-export type AccountLimitsQueryHookResult = ReturnType<typeof useAccountLimitsQuery>
-export type AccountLimitsLazyQueryHookResult = ReturnType<
-  typeof useAccountLimitsLazyQuery
->
-export type AccountLimitsQueryResult = Apollo.QueryResult<
-  AccountLimitsQuery,
-  AccountLimitsQueryVariables
 >
 export const WalletsDocument = gql`
   query wallets {
@@ -5625,6 +5290,218 @@ export type SendBitcoinDetailsScreenQueryResult = Apollo.QueryResult<
   SendBitcoinDetailsScreenQuery,
   SendBitcoinDetailsScreenQueryVariables
 >
+export const LnNoAmountInvoiceFeeProbeDocument = gql`
+  mutation lnNoAmountInvoiceFeeProbe($input: LnNoAmountInvoiceFeeProbeInput!) {
+    lnNoAmountInvoiceFeeProbe(input: $input) {
+      errors {
+        message
+      }
+      amount
+    }
+  }
+`
+export type LnNoAmountInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
+  LnNoAmountInvoiceFeeProbeMutation,
+  LnNoAmountInvoiceFeeProbeMutationVariables
+>
+
+/**
+ * __useLnNoAmountInvoiceFeeProbeMutation__
+ *
+ * To run a mutation, you first call `useLnNoAmountInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useLnNoAmountInvoiceFeeProbeMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [lnNoAmountInvoiceFeeProbeMutation, { data, loading, error }] = useLnNoAmountInvoiceFeeProbeMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useLnNoAmountInvoiceFeeProbeMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    LnNoAmountInvoiceFeeProbeMutation,
+    LnNoAmountInvoiceFeeProbeMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useMutation<
+    LnNoAmountInvoiceFeeProbeMutation,
+    LnNoAmountInvoiceFeeProbeMutationVariables
+  >(LnNoAmountInvoiceFeeProbeDocument, options)
+}
+export type LnNoAmountInvoiceFeeProbeMutationHookResult = ReturnType<
+  typeof useLnNoAmountInvoiceFeeProbeMutation
+>
+export type LnNoAmountInvoiceFeeProbeMutationResult =
+  Apollo.MutationResult<LnNoAmountInvoiceFeeProbeMutation>
+export type LnNoAmountInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
+  LnNoAmountInvoiceFeeProbeMutation,
+  LnNoAmountInvoiceFeeProbeMutationVariables
+>
+export const LnInvoiceFeeProbeDocument = gql`
+  mutation lnInvoiceFeeProbe($input: LnInvoiceFeeProbeInput!) {
+    lnInvoiceFeeProbe(input: $input) {
+      errors {
+        message
+      }
+      amount
+    }
+  }
+`
+export type LnInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
+  LnInvoiceFeeProbeMutation,
+  LnInvoiceFeeProbeMutationVariables
+>
+
+/**
+ * __useLnInvoiceFeeProbeMutation__
+ *
+ * To run a mutation, you first call `useLnInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useLnInvoiceFeeProbeMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [lnInvoiceFeeProbeMutation, { data, loading, error }] = useLnInvoiceFeeProbeMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useLnInvoiceFeeProbeMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    LnInvoiceFeeProbeMutation,
+    LnInvoiceFeeProbeMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useMutation<
+    LnInvoiceFeeProbeMutation,
+    LnInvoiceFeeProbeMutationVariables
+  >(LnInvoiceFeeProbeDocument, options)
+}
+export type LnInvoiceFeeProbeMutationHookResult = ReturnType<
+  typeof useLnInvoiceFeeProbeMutation
+>
+export type LnInvoiceFeeProbeMutationResult =
+  Apollo.MutationResult<LnInvoiceFeeProbeMutation>
+export type LnInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
+  LnInvoiceFeeProbeMutation,
+  LnInvoiceFeeProbeMutationVariables
+>
+export const LnUsdInvoiceFeeProbeDocument = gql`
+  mutation lnUsdInvoiceFeeProbe($input: LnUsdInvoiceFeeProbeInput!) {
+    lnUsdInvoiceFeeProbe(input: $input) {
+      errors {
+        message
+      }
+      amount
+    }
+  }
+`
+export type LnUsdInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
+  LnUsdInvoiceFeeProbeMutation,
+  LnUsdInvoiceFeeProbeMutationVariables
+>
+
+/**
+ * __useLnUsdInvoiceFeeProbeMutation__
+ *
+ * To run a mutation, you first call `useLnUsdInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useLnUsdInvoiceFeeProbeMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [lnUsdInvoiceFeeProbeMutation, { data, loading, error }] = useLnUsdInvoiceFeeProbeMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useLnUsdInvoiceFeeProbeMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    LnUsdInvoiceFeeProbeMutation,
+    LnUsdInvoiceFeeProbeMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useMutation<
+    LnUsdInvoiceFeeProbeMutation,
+    LnUsdInvoiceFeeProbeMutationVariables
+  >(LnUsdInvoiceFeeProbeDocument, options)
+}
+export type LnUsdInvoiceFeeProbeMutationHookResult = ReturnType<
+  typeof useLnUsdInvoiceFeeProbeMutation
+>
+export type LnUsdInvoiceFeeProbeMutationResult =
+  Apollo.MutationResult<LnUsdInvoiceFeeProbeMutation>
+export type LnUsdInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
+  LnUsdInvoiceFeeProbeMutation,
+  LnUsdInvoiceFeeProbeMutationVariables
+>
+export const LnNoAmountUsdInvoiceFeeProbeDocument = gql`
+  mutation lnNoAmountUsdInvoiceFeeProbe($input: LnNoAmountUsdInvoiceFeeProbeInput!) {
+    lnNoAmountUsdInvoiceFeeProbe(input: $input) {
+      errors {
+        message
+      }
+      amount
+    }
+  }
+`
+export type LnNoAmountUsdInvoiceFeeProbeMutationFn = Apollo.MutationFunction<
+  LnNoAmountUsdInvoiceFeeProbeMutation,
+  LnNoAmountUsdInvoiceFeeProbeMutationVariables
+>
+
+/**
+ * __useLnNoAmountUsdInvoiceFeeProbeMutation__
+ *
+ * To run a mutation, you first call `useLnNoAmountUsdInvoiceFeeProbeMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useLnNoAmountUsdInvoiceFeeProbeMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [lnNoAmountUsdInvoiceFeeProbeMutation, { data, loading, error }] = useLnNoAmountUsdInvoiceFeeProbeMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useLnNoAmountUsdInvoiceFeeProbeMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    LnNoAmountUsdInvoiceFeeProbeMutation,
+    LnNoAmountUsdInvoiceFeeProbeMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useMutation<
+    LnNoAmountUsdInvoiceFeeProbeMutation,
+    LnNoAmountUsdInvoiceFeeProbeMutationVariables
+  >(LnNoAmountUsdInvoiceFeeProbeDocument, options)
+}
+export type LnNoAmountUsdInvoiceFeeProbeMutationHookResult = ReturnType<
+  typeof useLnNoAmountUsdInvoiceFeeProbeMutation
+>
+export type LnNoAmountUsdInvoiceFeeProbeMutationResult =
+  Apollo.MutationResult<LnNoAmountUsdInvoiceFeeProbeMutation>
+export type LnNoAmountUsdInvoiceFeeProbeMutationOptions = Apollo.BaseMutationOptions<
+  LnNoAmountUsdInvoiceFeeProbeMutation,
+  LnNoAmountUsdInvoiceFeeProbeMutationVariables
+>
 export const AccountScreenDocument = gql`
   query AccountScreen {
     me {
@@ -5915,6 +5792,76 @@ export type SettingsScreenQueryResult = Apollo.QueryResult<
   SettingsScreenQuery,
   SettingsScreenQueryVariables
 >
+export const AccountLimitsDocument = gql`
+  query accountLimits {
+    me {
+      defaultAccount {
+        limits {
+          withdrawal {
+            totalLimit
+            remainingLimit
+            interval
+          }
+          internalSend {
+            totalLimit
+            remainingLimit
+            interval
+          }
+          convert {
+            totalLimit
+            remainingLimit
+            interval
+          }
+        }
+      }
+    }
+  }
+`
+
+/**
+ * __useAccountLimitsQuery__
+ *
+ * To run a query within a React component, call `useAccountLimitsQuery` and pass it any options that fit your needs.
+ * When your component renders, `useAccountLimitsQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useAccountLimitsQuery({
+ *   variables: {
+ *   },
+ * });
+ */
+export function useAccountLimitsQuery(
+  baseOptions?: Apollo.QueryHookOptions<AccountLimitsQuery, AccountLimitsQueryVariables>,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useQuery<AccountLimitsQuery, AccountLimitsQueryVariables>(
+    AccountLimitsDocument,
+    options,
+  )
+}
+export function useAccountLimitsLazyQuery(
+  baseOptions?: Apollo.LazyQueryHookOptions<
+    AccountLimitsQuery,
+    AccountLimitsQueryVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useLazyQuery<AccountLimitsQuery, AccountLimitsQueryVariables>(
+    AccountLimitsDocument,
+    options,
+  )
+}
+export type AccountLimitsQueryHookResult = ReturnType<typeof useAccountLimitsQuery>
+export type AccountLimitsLazyQueryHookResult = ReturnType<
+  typeof useAccountLimitsLazyQuery
+>
+export type AccountLimitsQueryResult = Apollo.QueryResult<
+  AccountLimitsQuery,
+  AccountLimitsQueryVariables
+>
 export const LocalizationContextProviderDocument = gql`
   query LocalizationContextProvider {
     me {
@@ -6018,3 +5965,56 @@ export function usePriceSubscription(
 }
 export type PriceSubscriptionHookResult = ReturnType<typeof usePriceSubscription>
 export type PriceSubscriptionResult = Apollo.SubscriptionResult<PriceSubscription>
+export const DeviceNotificationTokenCreateDocument = gql`
+  mutation deviceNotificationTokenCreate($input: DeviceNotificationTokenCreateInput!) {
+    deviceNotificationTokenCreate(input: $input) {
+      errors {
+        message
+      }
+      success
+    }
+  }
+`
+export type DeviceNotificationTokenCreateMutationFn = Apollo.MutationFunction<
+  DeviceNotificationTokenCreateMutation,
+  DeviceNotificationTokenCreateMutationVariables
+>
+
+/**
+ * __useDeviceNotificationTokenCreateMutation__
+ *
+ * To run a mutation, you first call `useDeviceNotificationTokenCreateMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useDeviceNotificationTokenCreateMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [deviceNotificationTokenCreateMutation, { data, loading, error }] = useDeviceNotificationTokenCreateMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useDeviceNotificationTokenCreateMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    DeviceNotificationTokenCreateMutation,
+    DeviceNotificationTokenCreateMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions }
+  return Apollo.useMutation<
+    DeviceNotificationTokenCreateMutation,
+    DeviceNotificationTokenCreateMutationVariables
+  >(DeviceNotificationTokenCreateDocument, options)
+}
+export type DeviceNotificationTokenCreateMutationHookResult = ReturnType<
+  typeof useDeviceNotificationTokenCreateMutation
+>
+export type DeviceNotificationTokenCreateMutationResult =
+  Apollo.MutationResult<DeviceNotificationTokenCreateMutation>
+export type DeviceNotificationTokenCreateMutationOptions = Apollo.BaseMutationOptions<
+  DeviceNotificationTokenCreateMutation,
+  DeviceNotificationTokenCreateMutationVariables
+>
