--- app/graphql/operations/remote/mutations.ts
@@ -1,48 +1,0 @@
-import { gql } from "@apollo/client"
-
-export default gql`
-  mutation deviceNotificationTokenCreate($input: DeviceNotificationTokenCreateInput!) {
-    deviceNotificationTokenCreate(input: $input) {
-      errors {
-        message
-      }
-      success
-    }
-  }
-
-  mutation lnNoAmountInvoiceFeeProbe($input: LnNoAmountInvoiceFeeProbeInput!) {
-    lnNoAmountInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-
-  mutation lnInvoiceFeeProbe($input: LnInvoiceFeeProbeInput!) {
-    lnInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-
-  mutation lnUsdInvoiceFeeProbe($input: LnUsdInvoiceFeeProbeInput!) {
-    lnUsdInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-
-  mutation lnNoAmountUsdInvoiceFeeProbe($input: LnNoAmountUsdInvoiceFeeProbeInput!) {
-    lnNoAmountUsdInvoiceFeeProbe(input: $input) {
-      errors {
-        message
-      }
-      amount
-    }
-  }
-`
